opsmedic_project/opsmedic_agent/requirements.txt
fastapi
uvicorn
docker
google-generativeai
requests
python-dotenv
openai